Move drawing to animation loader and add main menu scene
